feat: start to implement accordion ui for thememaps
